constructing UI

+++ exp_packages/SQLExecutor.py
@@ -4,6 +4,7 @@ import pymysql
 class SQLExecutor:
     db = None
     cursor = None
+    isConnected = False
 
     def __init__(self, server, username, passwd, database):
         self.server = server
@@ -18,34 +19,40 @@ class SQLExecutor:
                 self.server,
                 self.username,
                 self.passwd,
-                self.database)
+                self.database,
+                charset="utf8")
 
+            self.isConnected = True
             self.cursor = self.db.cursor()
+            self.cursor.execute("ALTER DATABASE {} CHARACTER SET utf8".format(self.database))
 
         except Exception as e:
-            print("SQLExecutor catched an error: {}".format(e))
+            print("SQLExecutor caught an error: {}".format(e))
             exit(1)
 
         print("[INFO] Database connected!")
 
     def close(self):
-        if self.db:
+        if self.isConnected:
             self.db.close()
+            self.isConnected = False
             print("[INFO] Database closed.")
 
     def create_database_table(self):
-        print("[INFO] Initializing Database Table...")
-        self.cursor.execute("DROP TABLE IF EXISTS StockData")
-        self.cursor.execute("CREATE TABLE StockData ("
-                            "StockCode int, "
-                            "StockName varchar(100),"
-                            "DataDate varchar(100),"
-                            "MarginBalance double(100,2),"
-                            "ClosingPrice double(100,2));")
+        if self.isConnected:
+            print("[INFO] Initializing Database Table...")
+            self.cursor.execute("DROP TABLE IF EXISTS StockData")
+            self.cursor.execute("CREATE TABLE StockData ("
+                                "StockCode int, "
+                                "StockName varchar(100),"
+                                "DataDate varchar(100),"
+                                "MarginBalance double(100,2),"
+                                "ClosingPrice double(100,2));")
+            self.cursor.execute("ALTER TABLE StockData CONVERT TO CHARACTER SET utf8")
+        else:
+            print("SQLExecutor caught an error: did not connect to database server")
 
     def exec(self, data):
-        print("[INFO] Inserting Fetched Data to Local...")
-
         try:
             self.cursor.executemany(
                 "INSERT INTO StockData(StockCode, StockName, DataDate, MarginBalance, ClosingPrice)VALUES(%s,%s,%s,%s,%s);",
@@ -53,3 +60,35 @@ class SQLExecutor:
 
         except Exception as e:
             print("SQLExecutor catched an error: {}".format(e))
+
+        self.db.commit()
+
+    def fetch(self, **kargs):
+        if self.isConnected:
+            query = "SELECT * FROM StockData WHERE 1"
+
+            for col, constraint in kargs.items():
+                query += " AND {} {}".format(col, constraint)
+
+            self.cursor.execute(query + ";")
+            result = self.cursor.fetchall()
+
+            if kargs["stockNameOnly"]:
+                ret = {}
+                return {result[i][0]: result[i][1]
+                        for i in range(len(result))
+                        if result[i][0] not in ret
+                        }
+
+            if kargs["timeIntervalOnly"]:
+                ret = []
+                return [result[i][2]
+                        for i in range(len(result))
+                        if result[i][2] not in ret
+                        ]
+
+            else:
+                return result
+
+        else:
+            print("SQLExecutor caught an error: did not connect to database server")
